Interfaces: Add job to checkConsistency call
we might need to know something about the job to check consistency
Add check to InstalledFiles.Exists function so that we check in cvmfs if necessary
avoids a bit of the hardcoded madness. Instead we could also completely pass on checking input files as is already done in most places.